copy output as test for minijava revisitor -> to rework to make proper tests

+++ tests/org.eclipse.emf.ecoretools.ale.compiler.interpreter.test/test-results/revisitor/logo_lang/revisitor-comp/test/impl/operation/impl/CosOpImpl.java
@@ -59,7 +59,7 @@ public class CosOpImpl extends UnaryExpressionOpImpl implements CosOp {
 
   public double eval(Turtle turtle) {
     double result;
-    result = TrigoServices.cosinus((double) (rev.$((Expression)this.obj.getExpression()).eval(((Turtle) (turtle)))));
+    result = ((double) (TrigoServices.cosinus((double) (rev.$((Expression)this.obj.getExpression()).eval(((Turtle) (turtle))))));
     return result;
   }
 }
