+++ tests/org.eclipse.emf.ecoretools.ale.compiler.interpreter.test/test-results/revisitor/logo_lang/revisitor-comp/test/impl/operation/impl/TanOpImpl.java
@@ -59,7 +59,7 @@ public class TanOpImpl extends UnaryExpressionOpImpl implements TanOp {
 
   public double eval(Turtle turtle) {
     double result;
-    result = TrigoServices.tan((double) (rev.$((Expression)this.obj.getExpression()).eval(((Turtle) (turtle)))));
+    result = ((double) (TrigoServices.tan((double) (rev.$((Expression)this.obj.getExpression()).eval(((Turtle) (turtle))))));
     return result;
   }
 }
